requirements.txt
pysendfile==2.0.1
nose==1.3.3
nose-testconfig==0.9.1
coverage==3.7.1